requirements.txt
streamlit
pdfplumber
docx2txt
spacy
pytesseract
Pillow
